[PRED-169] 보스 기본 로직 수정

+++ InWanderLand/LeftFrame.cpp
@@ -104,6 +104,11 @@ namespace BossSummon
 
 	void LeftFrame::SummonUnit()
 	{
+		if (!unitFrame.lock()->IsAlive())
+		{
+			return;
+		}
+
 		if (!summonCorountine.expired())
 		{
 			DeleteCoroutine(summonCorountine);
@@ -212,7 +217,7 @@ namespace BossSummon
 								auto targetUnit = PlayerController::Instance().GetPlayers().at(targetIndex - 1);
 								if (!targetUnit.expired() && targetUnit.lock()->IsAlive())
 								{
-									sUnit.lock()->OrderAttack(targetUnit);
+									sUnit.lock()->OrderAttackMove(targetUnit.lock()->GetTransform()->GetWorldPosition());
 									break;
 								}
 								findCount++;
@@ -245,7 +250,7 @@ namespace BossSummon
 								auto targetUnit = PlayerController::Instance().GetPlayers().at(targetIndex - 1);
 								if (!targetUnit.expired() && targetUnit.lock()->IsAlive())
 								{
-									sUnit.lock()->OrderAttack(targetUnit);
+									sUnit.lock()->OrderAttackMove(targetUnit.lock()->GetTransform()->GetWorldPosition());
 									break;
 								}
 								findCount++;
